restructure data connector folder
